add new file

+++ 123.py
@@ -1,4 +1,4 @@
 weight = input('請問你多重(kg)')
 height = input('請問你多高(cm)')
-print('你的BMI',float(weight)/(float(height)/100)**2)
+print('你的BMI',int(weight)/(int(height)/100)**2)
